pyfhevm/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.6
black==23.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
cytoolz==0.12.1
eip712-structs==1.1.0
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.2.0
frozenlist==1.3.3
hexbytes==0.3.1
idna==3.4
isort==5.12.0
jsonschema==4.18.1
jsonschema-specifications==2023.6.1
lru-dict==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
parsimonious==0.9.0
pathspec==0.11.1
platformdirs==3.8.1
protobuf==4.23.4
pycparser==2.21
pycryptodome==3.18.0
PyNaCl==1.5.0
pysha3==1.0.2
referencing==0.29.1
regex==2023.6.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.8.10
tomli==2.0.1
toolz==0.12.0
typing_extensions==4.7.1
urllib3==2.0.3
web3==6.5.0
websockets==11.0.3
yarl==1.9.2
